app.c: flush sendqueue on DTLS_EVENT_CONNECTED
When dtls_write() is called by send_to_peer(), it may not be able
to send all data immediately. This is particularly the case when
a DTLS (re-)handshake must be done first. When this happens, the
outstanding data is stored in a local sendqueue for this endpoint.

The contents of the sendqueue are flushed when the DTLS_EVENT_CONNECTED
event occured and before new data is sent to the same peer.